yasql/requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
Django==2.2.16
django-auth-ldap==2.2.0
django-celery-beat==1.5.0
django-celery-results==1.0.4
django-filter==2.3.0
django-redis==4.12.1
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
hiredis==1.1.0
humanfriendly==8.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
jdcal==1.4.1
kombu==4.6.11
msgpack==1.0.0
msgpack-python==0.5.6
mysql-replication==0.22
mysqlclient==2.0.1
openpyxl==3.0.5
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-ldap==3.3.1
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
zipp==3.2.0
zope.interface==5.1.0
